Setup for Combat and Animation architecture
Linked the blueprints for future gauntlet mesh to a shooting architecture and the character blueprint. Blueprint responses to button presses made by the player to shoot. Animation blueprint set up and connected to the character blueprint, awaiting a proper rigged mesh and animations. Blend Space also initialised for movement and awaiting animations